Implement ums scheduler creation on kernel module
Implement 'enter_ums_scheduling_mode' on kernel module which creates
a new ums scheduler or a new ums worker according to the passed input
flags.
The ums scheduler has a context associated with it that represents
its actual running state that can be restored when a ums worker
thread yields; it also has an event queue where ums scheduling events
are pushed. A reference to the newly created scheduler is stored
inside an idr which has the lifetime from the time the ums device is
opened to the time it is released.
Ums events are allocated by using a dedicated kmem cache reducing
possible bottlenecks on notifying new ums actions.
Dequeuing from the scheduler event queue can lead to sleeping if no
events are present in that moment, placing the task on the scheduler
wait queue.