Add functionality for switching UMS threads
Three new functionalities are implemented;
A UMS scheduler can execute a UMS worker thread, a UMS worker thread
can yield a result to a UMS scheduler and finally a UMS worker thread
can exit UMS mode.
Ums schedulers and ums workers are accessed through their ums context.
Each ums context is indexed inside a resizable hashtable with the key
corresponding to the virtual pid of the context task_struct.
The completion list lifecycle is managed by a kref: every scheduler and
every worker increments the kref counter when they are created and
decrements the kref counter when they are destroied.