Expose proc folders for showing ums statistics
A proc folder is created for each new process that is using UMS; Inside
that folder two subfolders are created: one folder containing each UMS
scheduler thread and one folder containing each UMS worker thread.
Every folder exposes an info file that shows some statistics as the
number of times an UMS thread switched to an other UMS thread and the
current state (idle/running).
An UMS scheduler proc folder also has a subfolder which contains a
symlink to every UMS worker thread that is in the scheduler completion
list.